+++ npm/src/directory-sync/batch-events/lock.ts
